Fix circular dependencies.
