+++ chatbot-core/tests/unit/rag/retrieve/test_retriever_utils.py
@@ -1,8 +1,9 @@
 """Unit Tests for retrieve_utils module."""
 
+import os
 import numpy as np
 import pytest
-from rag.retriever.retriever_utils import load_vector_index, search_index, INDEX_PATH, METADATA_PATH
+from rag.retriever.retriever_utils import load_vector_index, search_index, VECTOR_STORE_DIR
 
 def test_load_vector_index_returns_index_and_metadata(mocker):
     """Test load_vector_index loads index and metadata correctly."""
@@ -17,11 +18,15 @@ def test_load_vector_index_returns_index_and_metadata(mocker):
         "rag.retriever.retriever_utils.load_metadata",
         return_value=mock_metadata
     )
+    source_name = "plugins"
 
-    index, metadata = load_vector_index(mock_logger)
+    index, metadata = load_vector_index(mock_logger, source_name)
 
-    mock_load_index.assert_called_once_with(INDEX_PATH, mock_logger)
-    mock_load_metadata.assert_called_once_with(METADATA_PATH, mock_logger)
+    expected_index_path = os.path.join(VECTOR_STORE_DIR, f"{source_name}_index.idx")
+    expected_metadata_path = os.path.join(VECTOR_STORE_DIR, f"{source_name}_metadata.pkl")
+
+    mock_load_index.assert_called_once_with(expected_index_path, mock_logger)
+    mock_load_metadata.assert_called_once_with(expected_metadata_path, mock_logger)
     assert index == mock_index
     assert metadata == mock_metadata
